Add host tag only for `http://:9101` endpoint
To support customer-hosted prometheus, we need the `host` tag to be included in the prometheus client
that exports to `http://:9101`. Otherwise the metrics are not useful because they can't be differentiated
by host.

The VSOCK prometheus client filters out the `host` tag so that we comply with our security policies
regarding customer privacy.