Remove goreleaser releases
We need to build the amd/arm64 builds separately because the Telegraf
executable needs to be built with the right GOOS. Goreleaser does not
support releasing pre-built binaries in its free tier, which complicates
the release strategy. Instead of relying on Goreleaser to upload to
artifactory, we'll just upload directly ourselves and forego a proper
Github release. I'll need to think of something different (possibly pay
for the Pro tier of Goreleaser).

https://carlosbecker.com/posts/goreleaser-prebuilt/